Frontend: Changed Delete Button ("shoeDelete()") so now only shows when current User is creator of the post Backend: Extended delete function in PostController the delete request needs to be send by the author

+++ api/controllers/PostController.js
@@ -95,8 +95,14 @@ exports.delete = [
   authenticationRequired,
   rejectRequestsWithValidationErrors,
   (req, res) => {
-    PostModel.findByIdAndDelete(req.params.post_id)
-      .then(() => res.json('Post deleted.'))
+    PostModel.findById(req.params.post_id, 'author', function (err, post) {
+      if (post.author === req.user._id) {
+        PostModel.findByIdAndDelete(req.params.post_id)
+          .then(() => res.json('Post deleted.'))
+      } else {
+        res.status(401).json('Error: You do not have the right to delete this post!')
+      }
+    })
       .catch(err => res.status(400).json('Error: ' + err));
   }
 ];
